refactor(completion): import-sourced candidates gather through the CandidateSet
The imported (explicit qw names + remaining export surfaces) and
unimported (auto-import firehose) candidate sources move from the LSP
adapter into the set's complete() projection, tier-gated: explicit
imports are origin-file facts (OPEN), export surfaces are
dependency-file names (DEPENDENCY). The projection takes the slot's
import affordance (auto_import_at) — None reproduces the
suppress-firehose slots exactly; where the edit lands stays adapter
policy (auto_import_span). Presentation (details, priorities, edits
riding candidates) is unchanged.
